purger: do not fail when path does not exist
why: it can happen when there is an issue with celery task and file is
never created

+++ wazo_call_logd/purger.py
@@ -26,6 +26,10 @@ DEFAULT_CALL_LOGD_EXPORT_DAYS = 2
 
 def _remove_export_files(exports):
     for export in exports:
+        if not export.path:
+            logger.debug('Export file never created: "%s"', export.uuid)
+            continue
+
         try:
             os.remove(export.path)
         except FileNotFoundError:
